chore: migrated navButtonsConfig from NavBarContainer module into own config.tsx module
